requirements.txt

tensorflow-cpu>=2.12,<3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
